+++ js/inputmask.js
@@ -122,18 +122,18 @@
             importDataAttributes: true //import data-inputmask attributes
         },
         definitions: {
-            "9": {
-                validator: "[0-9]",
+            "9": { //\uFF11-\uFF19 #1606
+                validator: "[0-9\uFF11-\uFF19]",
                 cardinality: 1,
                 definitionSymbol: "*"
             },
-            "a": {
+            "a": { //\u0410-\u044F\u0401\u0451\u00C0-\u00FF\u00B5 #76
                 validator: "[A-Za-z\u0410-\u044F\u0401\u0451\u00C0-\u00FF\u00B5]",
                 cardinality: 1,
                 definitionSymbol: "*"
             },
             "*": {
-                validator: "[0-9A-Za-z\u0410-\u044F\u0401\u0451\u00C0-\u00FF\u00B5]",
+                validator: "[0-9\uFF11-\uFF19A-Za-z\u0410-\u044F\u0401\u0451\u00C0-\u00FF\u00B5]",
                 cardinality: 1
             }
         },